pyproject.toml
[tool.poetry]
name = "snappfood"
version = "0.1.0"
description = ""
authors = ["Naser <nrhosseinzadeh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
psycopg-binary = "^3.1.17"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.2.0"
psycopg = "^3.1.17"
celery = "^5.3.6"
pika = "^1.3.2"
redis = "^5.0.1"
hiredis = "^2.3.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
